# openapi/quickeval-v1.yaml
openapi: 3.1.0
info:
  title: QuickEval V1 API
  version: 0.1.0
  description: Internal API for the QuickEval Vue SPA.
servers:
  - url: /
...
paths:
...
  /api/v1/datasets:
    get:
      tags: [Datasets]
      operationId: listDatasets
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: evaluation_target_id
          in: query
          schema: { $ref: "#/components/schemas/UUID" }
        - name: scenario_id
          in: query
          schema: { $ref: "#/components/schemas/UUID" }
        - name: status
          in: query
          schema: { type: string, enum: [active, archived] }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Dataset page.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetPageResponse" }
    post:
      tags: [Datasets]
      operationId: createDataset
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DatasetRequest" }
      responses:
        "201":
          description: Dataset and its initial draft were created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateDatasetResponse" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/datasets/{dataset_id}:
    parameters:
      - $ref: "#/components/parameters/DatasetID"
    get:
      tags: [Datasets]
      operationId: getDataset
      responses:
        "200":
          description: Dataset and version history.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetDetailResponse" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [Datasets]
      operationId: updateDataset
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DatasetRequest" }
      responses:
        "200":
          description: Dataset updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetResponse" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    DatasetID:
      name: dataset_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: quickeval_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    UUID:
      type: string
      format: uuid
    UTCTimestamp:
      type: string
      format: date-time
      examples:
        - 2026-07-27T06:20:00.000Z
...
    DatasetRequest:
      type: object
      required: [scenario_id, name]
      properties:
        scenario_id: { $ref: "#/components/schemas/UUID" }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"] }
        expected_lock_version: { type: integer, minimum: 0, default: 0 }
...
    Dataset:
      type: object
      required:
        - id
        - scenario_id
        - scenario_name
        - evaluation_target_id
        - evaluation_target_name
        - name
        - status
        - lock_version
        - created_at
        - updated_at
        - published_version_count
        - draft_case_count
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        scenario_id: { $ref: "#/components/schemas/UUID" }
        scenario_name: { type: string }
        evaluation_target_id: { $ref: "#/components/schemas/UUID" }
        evaluation_target_name: { type: string }
        name: { type: string }
        description: { type: [string, "null"] }
        status: { type: string, enum: [active, archived] }
        lock_version: { type: integer, minimum: 0 }
        created_at: { $ref: "#/components/schemas/UTCTimestamp" }
        updated_at: { $ref: "#/components/schemas/UTCTimestamp" }
        latest_version_no: { type: [integer, "null"], minimum: 1 }
        published_version_count: { type: integer, minimum: 0 }
        draft_version_id: { type: [string, "null"], format: uuid }
        draft_case_count: { type: integer, minimum: 0 }
    DatasetVersion:
      type: object
      required:
        - id
        - dataset_id
        - status
        - lock_version
        - created_at
        - updated_at
        - case_count
        - enabled_count
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        dataset_id: { $ref: "#/components/schemas/UUID" }
        base_version_id: { type: [string, "null"], format: uuid }
        version_no: { type: [integer, "null"], minimum: 1 }
        status: { type: string, enum: [draft, published, archived] }
        release_note: { type: [string, "null"] }
        lock_version: { type: integer, minimum: 0 }
        published_at: { type: [string, "null"], format: date-time }
        archived_at: { type: [string, "null"], format: date-time }
        created_at: { $ref: "#/components/schemas/UTCTimestamp" }
        updated_at: { $ref: "#/components/schemas/UTCTimestamp" }
        case_count: { type: integer, minimum: 0 }
        enabled_count: { type: integer, minimum: 0 }
...
    DatasetResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Dataset" }
    CreateDatasetResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [dataset, draft]
              properties:
                dataset: { $ref: "#/components/schemas/Dataset" }
                draft: { $ref: "#/components/schemas/DatasetVersion" }
    DatasetDetailResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [dataset, versions]
              properties:
                dataset: { $ref: "#/components/schemas/Dataset" }
                versions:
                  type: array
                  items: { $ref: "#/components/schemas/DatasetVersion" }
...
    DatasetPageResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageBase"
                - type: object
                  properties:
                    items:
                      type: array
                      items: { $ref: "#/components/schemas/Dataset" }
...
    PageBase:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: {}
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    ResponseEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    ResponseMeta:
      type: object
      required:
        - request_id
      properties:
        request_id:
          type: string
          maxLength: 64
...
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
    Error:
      type: object
      required:
        - code
        - message
        - field_errors
      properties:
        code:
          type: string
        message:
          type: string
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required:
        - error
        - meta
      properties:
        error:
          $ref: "#/components/schemas/Error"
        meta:
          $ref: "#/components/schemas/ResponseMeta"